+++ src/routes/search/+page.svelte
@@ -1,21 +1,40 @@
 <script lang="ts">
     import Button from "../../components/Button.svelte";
     import Input from "../../components/Input.svelte";
-    import { ALL_VDs, type IVideo } from "../../lib/db";
+    import { ZPVDs, ZTVDs, type IVideo } from "../../lib/db";
     import { validate, normalize, pluralOrNot, similarity } from "strings-utils";
 
     let search = $state("");
-    let results: IVideo[] = $state([]);
+    let results: (IVideo & { similarity: number })[] = $state([]);
 
     $effect(() => {
         if (!validate(search) || search.length < 3) return;
 
-        results = ALL_VDs.filter(
-            (v) =>
-                normalize(v.title, { strict: true }).includes(search)
-                || normalize(v.topic, { strict: true }).includes(search)
-                || normalize(v.level, { strict: true }).includes(search)
-        ).sort((a, b) => similarity(a.title, search) - similarity(b.title, search));
+        // @ts-ignore
+        results = [...ZPVDs, ...ZTVDs]
+            .map((v) => {
+                const normalizedTitle = normalize(v.title, { strict: true });
+                const normalizedSubject = normalize(v.subject, { strict: true });
+                const normalizedLevel = normalize(v.level, { strict: true });
+
+                const similarityScore = search
+                    .toLowerCase()
+                    .split(/\s+/)
+                    .reduce((sum, word) => {
+                        // no sé como coño hacer que esto filtre por asignaturas
+                        // tipo que si contiene "mat", "mate" o cosas así no muestre cosas de física
+                        // ya lo pensaré, TODO
+                        const titleSim = similarity(normalizedTitle, word);
+                        const subjectSim = similarity(normalizedSubject, word);
+                        const levelSim = similarity(normalizedLevel, word);
+
+                        return sum + titleSim + subjectSim + levelSim;
+                    }, 0);
+
+                return { ...v, similarity: (similarityScore * 100) / 3 };
+            })
+            .filter((v) => v && v.similarity > 0)
+            .sort((a, b) => b!.similarity - a!.similarity);
     });
 </script>
 
@@ -52,39 +71,52 @@
         </p>
         <br />
         {#if results && results.length != 0}
-            {#each results as v}
-                <div
-                    class="flex w-full flex-col items-start justify-start gap-3 border-2 border-(--fff25) bg-(--blk) p-3 sm:flex-row"
-                >
-                    <img
-                        class="pointer-events-none! aspect-video w-full border-2 border-(--fff25) sm:max-w-60 lg:max-w-80"
-                        src={v.thumbnail}
-                        alt={`${v.title}, ${v.topic}, ${v.level}`}
-                    />
-                    <div class="flex flex-col gap-2">
-                        <h2>{v.title}</h2>
-                        <p>
-                            {v.topic} · {v.level} · <b>S{v.season}</b>
-                        </p>
-                        <br />
-                        <Button
-                            channel={v.channel}
-                            href={v.url}
-                            title="Redirigir a YouTube para ver este video."
-                            tail="w-auto mt-auto"
-                        >
-                            Ver en YouTube &gt;
-                        </Button>
-                        <p class="text-sm opacity-50">
-                            La S indica la temporada a la que pertenece el video; cada cambio
-                            significativo en la calidad de producción se sube el número,
-                            básicamente. Va desde el 0 (mis primeros videos) hasta el 4 (los más
-                            nuevos).
-                        </p>
+            <div class="grid grid-cols-1 gap-2 md:grid-cols-2 lg:grid-cols-4">
+                {#each results as v}
+                    <div class="flex flex-col gap-3 border-2 border-(--fff25) bg-(--blk) p-3">
+                        <img
+                            class="pointer-events-none! aspect-video w-full border-2 border-(--fff25)"
+                            src={v.thumbnail}
+                            alt={`${v.title}, ${v.subject}, ${v.level}`}
+                        />
+                        <div class="flex flex-1 flex-col justify-between gap-2">
+                            <div class="flex-10">
+                                <h2>{v.title}</h2>
+                                <p>
+                                    {v.subject} · {v.level} · <b>S{v.season}</b>
+                                </p>
+                            </div>
+                            <Button
+                                channel={v.channel}
+                                href={v.url}
+                                title="Redirigir a YouTube para ver este video."
+                                tail="h-auto mt-auto flex-1"
+                            >
+                                Ver en YouTube &gt;
+                            </Button>
+                            <p class="text-sm opacity-50">
+                                <b>{v.seen}</b> visualizaciones y <b>{v.likes}</b> «me gusta»<br
+                                />Se parece en un <b>{v.similarity.toFixed(2)}</b>% a lo que
+                                escribiste.
+                            </p>
+                        </div>
                     </div>
-                </div>
-                <br />
-            {/each}
+                {/each}
+            </div>
+            <br />
+            <p class="text-sm opacity-50">
+                La S indica la temporada a la que pertenece el video; cada cambio significativo en
+                la calidad de producción se sube el número, básicamente. Va desde el 0 (mis primeros
+                videos) hasta el 4 (los más nuevos y mejores).
+                <br /><br />
+                Los datos de visitas y «me gusta» no son en tiempo real, se actualizan aproximadamente
+                cada quince días (mucho tiempo para evitar consumo de red). Puedes ver lo más reciente
+                en YouTube, aunque no suelen cambiar con frecuencia.
+                <br /><br />
+                Recuerda que, en todos videos, el porcentaje de espectadores que expresan su opinión suele
+                ser bajo. 100 visitas y solo 15 «me gustas» no significan que a las otras 85 personas
+                no les haya gustado el video, no suele ser así.
+            </p>
         {:else}
             <p>Ningún resultado para <b>{search}</b>.</p>
         {/if}